init(05.09.2022): fix hw

+++ database/migrations/2022_08_30_120356_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('description');
             $table->int('category_id');
             $table->string('image');
-            $table->float('price');
+            $table->int('price');
             $table->boolean('active');
             $table->timestamps();
         });
